Added formik and created login/register page
I have added formik, then I have created auth components,
which display a little different content for login and register
page. It is based on the location: pathname props, then
we compair it with previously created array, if pathname
match to any of this item from array, this is passed as a prop
to the auth component.